moved tests to llama3_device folder
